Make ASL, LSR support updating memory or a register.
Unfortunately, I don't have a nice abstraction for the minor differences of reading or updating
the accumulator register vs a region of memory. The best I could do is factor out rmw_update and
special case the combination of a ReadModifyWrite instruction that uses Accumulator addressing.

We also needed some tweaks to how operands should be generated for Write instructions (sta,stx,sty)
as they were storing the region pointed to by the address rather than the address itself.